Create searcher.js
The purpose of this code is to look for a number in the participants of active conversation channels.
In some situations, a conversation session stays active or inactive rather than closed, resulting in the loss of future messages. 
If you are sending a WhatsApp message and it is not being delivered, use this function to see if a channel is still active.

This function is intended for use with conversations; this code will search the first 1000 conversation channels; 
this may be a limitation because it will only fetch the first 1000 elements; 
you can use a date filter to change the order of the conversations that will appear; 
try using the dates using the last time a message was successfully delivered to your Flex Instance.

The conversation sid will be returned in the output of this function. Use the function in line 63 to update the state to close. By default the function is commented.

Update the variable phoneNumber with the number you would like to search